refactoring: 약간의 최적화 및 변수명 변경

+++ aheui_core/src/cell.rs
@@ -1,6 +1,7 @@
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
+#[derive(Clone, Copy)]
 pub struct CellValue {
     pub position: Position,
     pub value: char
@@ -9,6 +10,6 @@ pub struct CellValue {
 #[wasm_bindgen]
 #[derive(Clone, Copy)]
 pub struct Position {
-    pub x: i8,
-    pub y: i8,
+    pub x: usize,
+    pub y: usize,
 }
